// src/app/(frontend)/reservation/types.ts
export type OperationResultType = {
  passenger: {
    modules: number[]
    passengers: {
      type: PassengerTypesIndexEnum
      gender: GenderEnumIndex
      fullName: string
      birthday: string
      identityNumber: string
      bookingCode: string
      campaignCode: null | string
      eTicketNumber: string
      firstName: string
      lastName: string
      mobilePhoneNumber: string
      email: string
      marketingAirlineCode: string
      isRoundedTrip: boolean
      module: number
      groupOrderIndex: number
      localPassengerSequenceNo: number
      localRelatedPassengerSequenceNo: null
      discount: {
        value: number
        currency: null
        rateValue: null
      }
      productItemId: ID
    }[]
    billingInformation: {
      billingName: string
      isCompany: boolean
      address: string
      taxNo: string
    }[]
    paymentInformation: {
      basketTotal: number
      basketDiscountTotal: number
      collectingTotal: number
      financellTotal: number
      mlTotal: null | number
      rateOfInterest: number
      installmentCount: number
      bankName: string
      encryptedCardHolder: string
      encryptedCardNumber: string
      sellingCurrency: 'TRY'
    }
    ssrList: []
    passengerCargoAddress: []
    bookingDateTime: string
    fromSession: boolean
    authorizeKey: null
    shoppingFileId: ID
    taxAmount: number
    shippingAmount: number
    operationResultPromotionUsageList: null
  }
  product: {
    summaryResponse: {
      moduleName:
        | 'Flight'
        | 'Hotel'
        | 'Bus'
        | 'Transfer'
        | 'CarRental'
        | 'Tour'
        | 'CyprusPackage'
    }
  }
  hotelCancelWarranty: {
    selectingCancelWarranty: boolean
    hasHotelWarranty: boolean
    cancelWarrantyPrice: number
    hotelWarrantyDiscountSelected: boolean
    totalPrice: number
    dayCount: number
    couponActive: boolean
  }
}